server/requirements.txt
asgiref==3.8.1
Django==4.2.16
django-filter==24.3
djangorestframework==3.15.2
django-cors-headers==4.4.0
importlib_metadata==8.5.0
Markdown==3.7
sqlparse==0.5.1
typing_extensions==4.12.2
yfinance==0.2.44
zipp==3.20.2
